Add allocation stopper and move stuff around
